+++ src/app/static/src/app/hintVersion.ts
@@ -1,1 +1,2 @@
-export const currentHintVersion = "1.36.0";
+export const currentHintVersion = "1.37.0";
+