+++ src/app/static/src/app/store/adrUpload/actions.ts
@@ -22,9 +22,9 @@ export const actions: ActionTree<ADRUploadState, RootState> & ADRUploadActions =
         const project = rootState.projects.currentProject;
 
         if (selectedDataset && project) {
-            context.commit({type: ADRMutation.SetADRError, payload: null});
+            commit({type: ADRUploadMutation.SetADRUploadError, payload: null});
             await api(context)
-                .withError(ADRMutation.SetADRError)
+                .withError(ADRUploadMutation.SetADRUploadError)
                 .ignoreSuccess()
                 .get(`/adr/datasets/${selectedDataset.id}`)
                 .then((response) => {
@@ -126,7 +126,7 @@ export const actions: ActionTree<ADRUploadState, RootState> & ADRUploadActions =
                 break
             }
         }
-        await dispatch("adr/getAndSetDatasets", selectedDatasetId);
-        dispatch("getUploadFiles");
+        await dispatch("adr/getAndSetDatasets", selectedDatasetId, {root: true});
+        await dispatch("getUploadFiles");
     }
 };